+++ example/main.dart
@@ -1,13 +1,13 @@
 import 'package:via_cep_flutter/via_cep_flutter.dart';
 
 Future<void> main() async {
-  final result = await readAddressByCep('05653-070');
+  final result = await readAddressByCep('31330-555');
 
-  print(result.city);
-  print(result.cep);
-  print(result.neighborhood);
-  print(result.state);
-  print(result.street);
+  if (result.isEmpty) return;
 
-  print(result.toJson());
+  print(result['city']);
+  print(result['cep']);
+  print(result['neighborhood']);
+  print(result['state']);
+  print(result['street']);
 }