+++ lib/services/via_cep.dart
@@ -6,7 +6,7 @@ import 'package:via_cep_flutter/error.dart';
 import 'package:via_cep_flutter/models/cep.dart';
 import 'package:http/http.dart' as http;
 
-Map<String, dynamic>? analyzeAndParseResponse(http.Response response) {
+Map<String, dynamic> analyzeAndParseResponse(http.Response response) {
   if (response.statusCode == 200) return json.decode(response.body);
 
   throw SimpleError('Erro ao se conectar com o serviço ViaCEP');
@@ -22,42 +22,35 @@ Map<String, dynamic> checkForViaCepError(Map<String, dynamic> responseObject) {
 
 Cep extractCepValuesFromResponse(Map<String, dynamic> responseObject) {
   return Cep(
-    cep: responseObject['cep'].replaceFirst('-', ''),
-    city: responseObject['localidade'],
-    neighborhood: responseObject['bairro'],
-    state: responseObject['uf'],
-    street: responseObject['logradouro'],
-  );
-}
-
-void throwApplicationError(Object e) {
-  final String? message = e is SimpleError
-      ? e.message
-      : 'Erro ao se conectar com o serviço ViaCEP.';
-
-  throw ServiceError(
-    service: Service.ViaCEP,
-    message: message,
+    responseObject['cep'].replaceFirst('-', ''),
+    responseObject['uf'],
+    responseObject['localidade'],
+    responseObject['logradouro'],
+    responseObject['bairro'],
   );
 }
 
 Future<Cep> fetchViaCepService(String cepWithLeftPad) async {
   final url = Uri.parse('https://viacep.com.br/ws/$cepWithLeftPad/json/');
-  Cep resposta = const Cep();
-
-  final response = await http.get(
-    url,
-    headers: {
-      'content-type': 'application/json;charset=utf-8',
-    },
-  );
-
+  late Cep resposta;
   try {
-    final analyze = analyzeAndParseResponse(response);
-
-    if (analyze != null) resposta = extractCepValuesFromResponse(analyze);
+    final httpresponse = await http.get(
+      url,
+      headers: {
+        'content-type': 'application/json;charset=utf-8',
+      },
+    );
+    final analyze = analyzeAndParseResponse(httpresponse);
+    resposta = extractCepValuesFromResponse(analyze);
   } catch (e) {
-    throwApplicationError(e);
+    final String? message = e is SimpleError
+        ? e.message
+        : 'Erro ao se conectar com o serviço ViaCEP.';
+
+    throw ServiceError(
+      service: Service.ViaCEP,
+      message: message,
+    );
   }
 
   return resposta;