+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGES
 
+## 0.2.2
+
+ * Added more text to the ``` README.md ``` file, telling how to handle possible errors.
+ * Added more documentation to the code 
+ * Fixed ```UNHANDLE EXCEPTION``` bug. Now even if the Cep is invalid, the code will still run.
+ * Now instead of return a ```Cep```, ```readAddressByCep``` returns a ```Map<String, dynamic>```
+
+ * Some other changes to the code...
+
 ## 0.1.2 
 
  * Removed text from the ```README.md``` file that was from the original repository